[fix]: modified the ODT document parser class to communicate with a common interface
