+++ src/odt/examples/parsing_nodes.py
@@ -1,15 +1,12 @@
 from src.helpers.odt import consts
 from src.odt.elements.ODTDocument import ODTDocument
 from src.odt.ODTParser import ODTParser
-from src.odt.elements.StylesContainer import StylesContainer
 
 if __name__ == '__main__':
-    doc_path = "documents/tabl1.odt"
+    doc_path = "/Users/vladtereshch/PycharmProjects/normcontrol-Document-Parser/src/odt/documents/tabl1.odt"
     doc = ODTDocument(doc_path)
-    odt_parser = ODTParser()
-    styles_container = StylesContainer(doc)
+    odt_parser = ODTParser(doc)
 
-    styles_container.build_dict()
-    list = styles_container.get_nodes_with_style_full7(doc.document.text, consts.DEFAULT_PARAM)
+    list = odt_parser.get_nodes_with_style_full7(doc.document.text, consts.DEFAULT_PARAM)
     print(list)
     print(list["0"])
