+++ src/odt/examples/parsing_styles.py
@@ -1,11 +1,10 @@
 from src.odt.elements.ODTDocument import ODTDocument
 from src.odt.ODTParser import ODTParser
-from src.odt.elements.StylesContainer import StylesContainer
 
 if __name__ == '__main__':
-    doc_path = "documents/listsimages.odt"
+    doc_path = "/Users/vladtereshch/PycharmProjects/normcontrol-Document-Parser/src/odt/documents/listsimages.odt"
     doc = ODTDocument(doc_path)
-    odt_parser = ODTParser()
+    odt_parser = ODTParser(doc)
 
     print("Получение текста и автоматических стилей:\n")
     print(odt_parser.automatic_style_parser.get_automatic_styles(doc))
@@ -40,7 +39,6 @@ if __name__ == '__main__':
     print("-------------------------------------------\n")
 
     print("Печать всех стилей документа:")
-    styles_container = StylesContainer(doc)
-    print(styles_container.all_automatic_styles)
-    print(styles_container.all_default_styles)
-    print(styles_container.all_regular_styles)
+    print(odt_parser.all_automatic_styles)
+    print(odt_parser.all_default_styles)
+    print(odt_parser.all_regular_styles)
