+++ src/odt/examples/parsing_tables.py
@@ -1,12 +1,11 @@
 from src.helpers.odt import consts
 from src.odt.elements.ODTDocument import ODTDocument
 from src.odt.ODTParser import ODTParser
-from src.odt.elements.StylesContainer import StylesContainer
 
 if __name__ == '__main__':
-    doc_path = "documents/tabl1.odt"
+    doc_path = "/Users/vladtereshch/PycharmProjects/normcontrol-Document-Parser/src/odt/documents/tabl1.odt"
     doc = ODTDocument(doc_path)
-    odt_parser = ODTParser()
+    odt_parser = ODTParser(doc)
 
     print("Получение текста и автоматических стилей:\n")
     print(odt_parser.automatic_style_parser.get_automatic_styles(doc))
@@ -45,9 +44,6 @@ if __name__ == '__main__':
     print(odt_parser.table_parser.get_table_column_parameter(ast, 'column-width'))
     print("-------------------------------------------\n")
 
-    styles_container = StylesContainer(doc)
-
-    styles_container.build_dict()
-    all_doc_info = styles_container.get_nodes_with_style_full7(doc.document.text, consts.DEFAULT_PARAM)
+    all_doc_info = odt_parser.get_nodes_with_style_full7(doc.document.text, consts.DEFAULT_PARAM)
     auto_pars = odt_parser.paragraph_parser.paragraphs_helper(all_doc_info)
     print(auto_pars)
